fix empty shopping cart bug

+++ cse305_final/main.py
@@ -373,19 +373,22 @@ def checkout():
         for item in items:
             totalPrice += item[2]
             string += item[1] + " "
-        try:
-            cur.execute("INSERT INTO Orders(CustomerID, OrderID, Items, TotalPrice, PlacedOn) VALUES (?, ?, ?, ?, ?)", (customer, orderID, string, totalPrice, now))
-            cur.execute("DELETE FROM ShoppingCart WHERE CustomerID = ?", (customer, ))
-            edb.commit()
-            # cur.execute("SELECT ArticleID FROM ShoppincCart WHERE CustomerID = ?", (customer, ))
-            # items = cur.fetchall()
-            output = "Order successfully placed"
-        except:
-            edb.rollback()
-            output = "Could not place order"
-    edb.close()
-    print(output)
-    return render_template("checkout.html", orderID=orderID, now=now, name=FirstName, address=address)
+        if totalPrice != 0:
+            try:
+                cur.execute("INSERT INTO Orders(CustomerID, OrderID, Items, TotalPrice, PlacedOn) VALUES (?, ?, ?, ?, ?)", (customer, orderID, string, totalPrice, now))
+                cur.execute("DELETE FROM ShoppingCart WHERE CustomerID = ?", (customer, ))
+                edb.commit()
+                output = "Order successfully placed"
+            except:
+                edb.rollback()
+                output = "Could not place order"
+            edb.close()
+            print(output)
+            return render_template("checkout.html", orderID=orderID, now=now, name=FirstName, address=address, totalPrice=totalPrice)
+        else:
+            edb.close()
+            return redirect(url_for('/ShoppingCart'))
+
 
 @app.route("/logout")
 def logout():
